+++ CHANGELOG.md
@@ -12,6 +12,8 @@ All notable changes to the Aptos TypeScript SDK will be captured in this file. T
 - [`Fix`] `getModuleEventsByEventType` will also account for EventHandle events.
 - [`Hot Fix`] change regex to find object address when using `createObjectAndPublishPackage` move function
 - Add support for federated keyless accounts as defined in AIP-96
+- Add a `signAndSubmitAsFeePayer` function to the API.
+- Add an optional `feePayerAuthenticator` parameter to `signAndSubmitTransaction`.
 
 # 1.27.1 (2024-08-23)